WebStorage/tests/submissions/Infraware/storage_session_removeitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_removeItem</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();
			sessionStorage.setItem("name", "user1");

			if (sessionStorage.getItem("name") == "user1")
			{
				try {
					sessionStorage.removeItem("name");
					sessionStorage.removeItem("unknown");
				} catch (e) {
					test(function() {
						assert_unreached("removeItem method is failed.");
					}, "removeItem test")
				}

				test(function() {
					assert_equals(sessionStorage.getItem("name"), null, "sessionStorage.getItem('name')")
				}, "result test of removeItem method")

			} else {
				test(function() {
					assert_unreached("'name' item is not exist.");
				}, "item existence test")
			}
		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
